requirements.txt
fastapi==0.135.1
langchain_community==0.4.1
langchain_core==1.2.20
langchain_openai==1.1.11
langchain_qdrant==1.1.0
langchain_text_splitters==1.1.1
pydantic==2.12.5
python-dotenv==1.2.2
qdrant_client==1.17.1
Requests==2.32.5
